+++ cpp/CMakeLists.txt
@@ -109,8 +109,8 @@ endif()
 # ##################################################################################################
 # * library targets --------------------------------------------------------------------------------
 
-file(GLOB SOURCES "src/*.cpp" "src/buffer/*.cpp" "src/communicator/*.cpp" "src/cudf/*.cpp"
-     "src/shuffler/*.cpp"
+file(GLOB SOURCES "src/*.cpp" "src/buffer/*.cpp" "src/communicator/*.cpp"
+     "src/integrations/cudf/*.cpp" "src/shuffler/*.cpp"
 )
 
 # Without MPI support, remove the MPI communicator from the source list.
