+++ python/rapidsmpf/rapidsmpf/examples/bulk_mpi_shuffle.py
@@ -20,7 +20,7 @@ import rapidsmpf.communicator.mpi
 from rapidsmpf.buffer.buffer import MemoryType
 from rapidsmpf.buffer.resource import BufferResource, LimitAvailableMemory
 from rapidsmpf.config import Options, get_environment_variables
-from rapidsmpf.cudf.partition import (
+from rapidsmpf.integrations.cudf.partition import (
     partition_and_pack,
     unpack_and_concat,
 )
