+++ python/rapidsmpf/rapidsmpf/tests/test_shuffler.py
@@ -12,7 +12,7 @@ import cudf
 from rmm.pylibrmm.stream import DEFAULT_STREAM
 
 from rapidsmpf.buffer.resource import BufferResource
-from rapidsmpf.cudf.partition import (
+from rapidsmpf.integrations.cudf.partition import (
     partition_and_pack,
     split_and_pack,
     unpack_and_concat,
